refactor existing code structure
