+++ packages/rlc-codegen/src/index.ts
@@ -13,14 +13,14 @@ export * from "./buildParameterTypes.js";
 export * from "./buildIsUnexpectedHelper.js";
 export * from "./buildTopLevelIndexFile.js";
 export * from "./buildIndexFile.js";
-export * from "./buildPackageFile.js";
-export * from "./buildRollupConfig.js";
-export * from "./buildTsConfig.js";
-export * from "./buildApiExtractorConfig.js";
 export * from "./buildPaginateHelper.js";
 export * from "./buildPollingHelper.js";
-export * from "./buildESLintConfig.js";
 export * from "./test/buildKarmaConfig.js";
 export * from "./test/buildEnvFile.js";
 export * from "./test/buildRecordedClient.js";
 export * from "./test/buildSampleTest.js";
+export * from "./metadata/buildApiExtractorConfig.js";
+export * from "./metadata/buildPackageFile.js";
+export * from "./metadata/buildRollupConfig.js";
+export * from "./metadata/buildTsConfig.js";
+export * from "./metadata/buildESLintConfig.js";
